Former-commit-id: bdfd408aeb870e69a0cf6dee896f40b9a5683c6b

+++ ifs-web-service/ifs-competition-mgt-service/src/main/java/org/innovateuk/ifs/competitionsetup/service/sectionupdaters/AdditionalInfoSectionSaver.java
@@ -14,11 +14,11 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.Optional;
 
-import static org.codehaus.groovy.runtime.InvokerHelper.asList;
+import static java.util.Arrays.asList;
 import static org.innovateuk.ifs.commons.service.ServiceResult.serviceFailure;
 import static org.innovateuk.ifs.commons.service.ServiceResult.serviceSuccess;
 
@@ -114,7 +114,7 @@ public class AdditionalInfoSectionSaver extends AbstractSectionSaver implements 
 		if (fieldName.endsWith("funder")) {
 			funder.setFunder(value);
 		} else if(fieldName.endsWith("funderBudget")) {
-			funder.setFunderBudget(new BigDecimal(value));
+			funder.setFunderBudget(new BigInteger(value));
 		} else {
 			return serviceFailure(new Error("Field not found", HttpStatus.BAD_REQUEST));
 		}
